fixL MET-1311 token link in address and stake key detail
